Keep it in sync with jenkins_node.yml
Right now Plone jenkins nodes are configured via jenkins_node.yml
config file that is found on jenkins.plone.org repo.

This playbook is a role based on that configuration with some
extra candies.

Try to keep them as in sync as possible until a decision is made
on which one should prevail:
https://github.com/plone/jenkins.plone.org/issues/104