Fixed installing SSH public key
When ansible does a lookup (on the next step "user: install authorized key")
it does on the local computer not the remote server you are running
ansible commands on.

This means that to install an SSH public key you have to download
the public key on your local computer and then the lookup will work.